renaming v3 to to static
